reworked the Color threshold sliders
the main.dart was starting to become way to long so we decide to start
the work to breakdown the main Page into smaller widgets.

first one is the Sliders responsible to change the threshold of the RGB
channels.

so now we have the Sliders being used into two smaller layers:

- ColorSlider: help us abstract the SliderTheme and the necessary
  callback when changing values.

- BoundLevelColorSliders: help us abstract the setup of using three
  ColorSlider widgets with a single "title". this is more or less a sort
  of proxy/factory/(facade?).

that also triggered a small rework on the folder names as well (nothing
fancy).